text-to-music/requirements.txt
transformers>=4.30.0
torch>=2.0.0
numpy>=1.21.0
librosa>=0.9.0
soundfile>=0.10.3
pydub>=0.25.1
python-dotenv>=0.19.0
requests>=2.26.0
tqdm>=4.62.0
scipy>=1.7.0
dashscope>=1.10.0  # 通义千问API SDK 